dom/browser-element/mochitest/priority/test_ExpectingSystemMessagePriority2.html

<!DOCTYPE HTML>
<html>
<!--
Test that a regular (not mozapptype=critical) frame that's expecting a system
message gets priority BACKGROUND_PERCEIVABLE when it's in the background.
-->
<head>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="../browserElementTestHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.7">
"use strict";

// Basic browser element test settings.
SimpleTest.waitForExplicitFinish();
browserElementTestHelpers.setEnabledPref(true);
browserElementTestHelpers.addPermission();
browserElementTestHelpers.enableProcessPriorityManager();

var basePath = 'http://mochi.test:8888/tests/dom/browser-element/mochitest/';
var gHostedManifestURL = basePath + 'file_app.sjs?testToken=priority/file_SystemMessage.html';
var gHostedPageURL = basePath +'priority/file_SystemMessage.html';
var gApp;

var testingMessage = 'Test expecting system message!';

addEventListener('testready', function() {
  function cbError() {
    ok(false, "Error callback invoked " + this.error.name);
    SimpleTest.finish();
  }

  function installApp() {
    var request = navigator.mozApps.install(gHostedManifestURL);
    request.onerror = cbError;
    request.onsuccess = function() {
      gApp = request.result;
      runTest();
    }
  }

  function uninstallApp() {
    // Uninstall the app.
    var request = navigator.mozApps.mgmt.uninstall(gApp);
    request.onerror = cbError;
    request.onsuccess = function() {
      // All done.
      info("All done");
      runTest();
    }
  }

  // ProcessPriorityManager requires at least one process in foreground
  // so that other processes can transit freely between foreground and
  // background.
  function setupTest() {
    var foreground = document.createElement('iframe');
    foreground.setAttribute('mozbrowser', true);
    foreground.src = browserElementTestHelpers.emptyPage;
    expectMozbrowserEvent(foreground, 'loadend').then(runTest);
    document.body.appendChild(foreground);
  }

  function testApp() {

    var iframe = document.createElement('iframe');
    iframe.setAttribute('mozbrowser', true);
    iframe.setAttribute('mozapp', gApp.manifestURL);
    iframe.setAttribute('src', gApp.manifest.launch_path);
    //iframe.setAttribute('src', browserElementTestHelpers.emptyPage);

    var childID = null;

    expectProcessCreated('FOREGROUND').then(function(chid) {
      childID = chid;
      info('process created with childID ' + childID);
    }).then(function() {
      info('waiting page loaded...');
      return expectMozbrowserEvent(iframe, 'loadend');
    }).then(function() {
      info('switch this frame to background, and expected priority change.');
      var p = expectPriorityChange(childID, 'BACKGROUND_PERCEIVABLE');
      iframe.setVisible(false);
      return p;
    }).then(function() {
      info('message time out, expect process priority changed.');
      return expectPriorityChange(childID, 'BACKGROUND');
    }).then(SimpleTest.finish);

    document.body.appendChild(iframe);
  }

  var tests = [
    // Permissions
    function() {
      SpecialPowers.pushPermissions(
        [{ "type": "embed-apps", "allow": 1, "context": document },
         { "type": "webapps-manage", "allow": 1, "context": document }], runTest);
    },

    // Preferences.
    function() {
      SpecialPowers.pushPrefEnv({"set": [["dom.ipc.systemMessageCPULockTimeoutSec", 5],
                                         ["dom.sysmsg.enabled", true],
                                         ["dom.testing.ignore_ipc_principal", true]]}, runTest);
    },

    function() {
      SpecialPowers.setAllAppsLaunchable(true);
      // No confirmation needed when an app is installed.
      SpecialPowers.autoConfirmAppInstall(() => {
        SpecialPowers.autoConfirmAppUninstall(runTest);
      });
    },

    function() {
      if (SpecialPowers.isMainProcess()) {
        SpecialPowers.Cu.import("resource://gre/modules/Services.jsm");
        var ioService = SpecialPowers.Cc["@mozilla.org/network/io-service;1"]
                                     .getService(SpecialPowers.Ci.nsIIOService);
        var systemMessenger = SpecialPowers.Cc["@mozilla.org/system-message-internal;1"]
                              .getService(SpecialPowers.Ci.nsISystemMessagesInternal);
        systemMessenger.registerPage("request-sync",
                                     ioService.newURI(gHostedPageURL, null, null),
                                     ioService.newURI(gHostedManifestURL, null, null));
        systemMessenger.broadcastMessage("request-sync", {msg: testingMessage});
      }
      runTest();
    },

    // Installing the app.
    installApp,

    // Insert one mozbrowser frame as the foreground frame
    setupTest,

    // Run tests in app.
    testApp,

    // Uninstall the app.
    uninstallApp
  ];

  function runTest() {
    if (!tests.length) {
      SimpleTest.finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  runTest();

});

</script>
</body>
</html>
